Light sensor program
This program tells the device to sleep until light over a set intensity is measured. Once the intensity threshold is met, the device records the intensity value and logs it. Intended for use in caves but could be modified for any sort of light detection (e.g. monitoring a fridge/garage/etc opening and closing)